Separate sort functions into new utilty file
